Add files via upload
This file shows the size and time differences between the old main functions and the new main functions that are used to simulate a model. The times for these functions are based off of simulating tract number 030400 and county number 087 of the state of Louisiana (number 22) fifty times. The minimum, maximum, and mean time from the fifty trials for each main function is recorded. It is important to note that the size of this tract was not very big, so the time it takes to simulate other tracts could be significantly different (potentially 5 times or more longer than the recorded times) depending on a tract's size.