Add files via upload
This file shows the differences in size and time for all the helper functions that are called from the household_generator() and group_quarters_simulator() functions. The times for the first thirteen functions were found by testing how long it took each function to simulate the characteristics for all 413 households in tract number 030400 and county number 087 of Louisiana (state number 22). This was done in trials of fifty, with the minimum, maximum, and mean time from the fifty trials being recorded. The times for the last three functions were found by testing how long it took each function to simulate the characteristics for all 287 group quarters in the same tract, county, and state as the households. This was also done in fifty trials. The number of households and group quarters for this tract was not very large, so the times for other tracts could be very different depending on the tract's size.