Update new_groupquarterssimulator.R
Now almost all the characteristics for an individual are simulated at once (in the getindividualcharacteristics() function instead of one at a time. This reduces the number of times the code originally had to loop through the same individuals to simulate all the characteristics. The run time is now approximately halved.